test_requirements.txt
importlib
djangocms-helper
mock
pysqlite
coverage
